Helping with corrupted DETS files.
The corrupted files happen due to {bad_object,Reason} errors.
The two common errors reported were 'read_buckets' and
'eval_work_list', which seem to be related to bugs within
DETS and files closed incorrectly.

There doesn't seem to be much to be done to solve the issue.
Although copy/pasting the file somewhere else and changing its
name seemed to work fine sometimes, it didn't always do so.

The new code warns the user in case of problem, avoids the
infinite crashing that was previously there, and attempts to
repair the DETS table (usually unsuccessfully).

Attempt to resolve issue #2