Allowing multiple independent versions.
This helps handling changes happening with the main
OTP releases. Different files will be available for all
different releases.

The make file has been fixed to allow automatic building based
on the version and the documentation has been updated accordingly.

This should fix issue #3, where the thing was broken for R13B04.
Thanks to Radosław Szymczyszyn (@lavrin) for the fixes on that one.