Reinstall the files if possible, just skip backups
The current behaviour stops the install the moment a backup file of
group.erl is found.

However, if the user wants to reinstall multiple times, it might be the
sign of a patch. It is not a problem to install multiple times, only the
original backup of group.erl must be preserved.

This patch moves the install code outside of the backup checking path to
allow for this.